Fix bug in texture sample creation
The creation of a texture sample would fail when done internally
through the new-from-template mechanims (see
https://numpy.org/doc/stable/user/basics.subclassing.html). This
wasn't spotted earlier because that's not something I had ever used or
tested. But apparently now this mechanism is used when calling
imshow (so some change must have occurred in recent versions of
matplotlib), so this would result in an exception when trying to
plot images.

Fixes issue #1.